Sign in controller logic should be the same for Twitter and Mastodon OAuth2; so enforce with with an common class.
